Break down Example into its own Service
