+++ newIDE/app/src/ProjectCreation/LocalExamples.js
@@ -4,11 +4,11 @@ import { I18n } from '@lingui/react';
 import { type I18n as I18nType } from '@lingui/core';
 import * as React from 'react';
 import { ExampleStore } from '../AssetStore/ExampleStore';
-import { getExample } from '../Utils/GDevelopServices/Asset';
+import { getExample } from '../Utils/GDevelopServices/Example';
 import Divider from '@material-ui/core/Divider';
 import LocalFolderPicker from '../UI/LocalFolderPicker';
 import { sendNewGameCreated } from '../Utils/Analytics/EventSender';
-import { type ExampleShortHeader } from '../Utils/GDevelopServices/Asset';
+import { type ExampleShortHeader } from '../Utils/GDevelopServices/Example';
 import { Column, Line } from '../UI/Grid';
 import optionalRequire from '../Utils/OptionalRequire.js';
 import { showErrorBox } from '../UI/Messages/MessageBox';
