+++ newIDE/app/src/Utils/GDevelopServices/Asset.js
@@ -3,7 +3,6 @@ import axios from 'axios';
 import { GDevelopAssetApi } from './ApiConfigs';
 import semverSatisfies from 'semver/functions/satisfies';
 import { type Filters } from './Types.flow';
-import { type UserPublicProfile } from './User';
 
 export type SerializedParameterMetadata = {|
   codeOnly: boolean,
@@ -81,29 +80,6 @@ export type AllResources = {|
   filters: Filters,
 |};
 
-export type ExampleShortHeader = {|
-  id: string,
-  name: string,
-  shortDescription: string,
-  license: string,
-  tags: Array<string>,
-  authors?: Array<UserPublicProfile>,
-  previewImageUrls: Array<string>,
-  gdevelopVersion: string,
-|};
-
-export type Example = {|
-  ...ExampleShortHeader,
-  description: string,
-  projectFileUrl: string,
-  authors: Array<string>,
-|};
-
-export type AllExamples = {|
-  exampleShortHeaders: Array<ExampleShortHeader>,
-  filters: Filters,
-|};
-
 export type License = {|
   name: string,
   website: string,
@@ -155,40 +131,6 @@ export const getAsset = (
       }
 
       return axios.get(assetUrl);
-    })
-    .then(response => response.data);
-};
-
-export const listAllExamples = (): Promise<AllExamples> => {
-  return axios
-    .get(`${GDevelopAssetApi.baseUrl}/example`)
-    .then(response => response.data)
-    .then(({ exampleShortHeadersUrl, filtersUrl }) => {
-      if (!exampleShortHeadersUrl || !filtersUrl) {
-        throw new Error('Unexpected response from the example endpoint.');
-      }
-      return Promise.all([
-        axios.get(exampleShortHeadersUrl).then(response => response.data),
-        axios.get(filtersUrl).then(response => response.data),
-      ]).then(([exampleShortHeaders, filters]) => ({
-        exampleShortHeaders,
-        filters,
-      }));
-    });
-};
-
-export const getExample = (
-  exampleShortHeader: ExampleShortHeader
-): Promise<Example> => {
-  return axios
-    .get(`${GDevelopAssetApi.baseUrl}/example/${exampleShortHeader.id}`)
-    .then(response => response.data)
-    .then(({ exampleUrl }) => {
-      if (!exampleUrl) {
-        throw new Error('Unexpected response from the example endpoint.');
-      }
-
-      return axios.get(exampleUrl);
     })
     .then(response => response.data);
 };
